+++ spec/src/options_behavior.spec.ts
@@ -58,43 +58,78 @@ describe('options behavior', () => {
 
         describe('when value is null and option is null/undefined/false (default)', () => {
 
+            class Doctor {
+                @Serialize() public name: string;
+            }
+
             class Patient {
-                @Serialize() public married: boolean;
-                public constructor(married?: boolean) {
-                    if (married !== undefined) { this.married = married; }
-                }
+                @Serialize() public age: number;
+                @Serialize() public doctor: Doctor;
             }
 
             it('should fail to serialize', () => {
-                const patient = new Patient(null);
-                expect(() => deflate(patient)).to.throw('Unable to serialize property "married": Value is null');
+
+                const doctor = Object.assign(new Doctor(), { name: null });
+                expect(
+                    () => deflate(doctor)
+                ).to.throw('Unable to serialize property "name": Value is null');
+
+                const patient = Object.assign(new Patient(), { age: 35, doctor: null });
+                expect(
+                    () => deflate(patient)
+                ).to.throw('Unable to serialize property "doctor": Value is null');
+
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Patient, { married: null })).to.throw('Unable to deserialize property "married": Value is null');
+
+                expect(
+                    () => inflate(Doctor, { name: null })
+                ).to.throw('Unable to deserialize property "name": Value is null');
+
+                expect(
+                    () => inflate(Patient, { age: 35, doctor: null })
+                ).to.throw('Unable to deserialize property "doctor": Value is null');
+
             });
 
         });
 
         describe('when value is null and option is true', () => {
 
+            class Doctor {
+                @Serialize({ nullable: true }) public name: string;
+            }
+
             class Patient {
-                @Serialize({ nullable: true }) public married: boolean;
-                public constructor(married?: boolean) {
-                    if (married !== undefined) { this.married = married; }
-                }
+                @Serialize({ nullable: true }) public age: number;
+                @Serialize({ nullable: true }) public doctor: Doctor;
             }
 
             it('serializes to null', () => {
-                const patient = new Patient(null);
-                const serialized = deflate(patient);
-                expect(serialized).to.deep.equal({ married: null });
+
+                const doctor = Object.assign(new Doctor(), { name: null });
+                expect(
+                    deflate(doctor)
+                ).to.deep.equal({ name: null });
+
+                const patient = Object.assign(new Patient(), { age: 35, doctor: null });
+                expect(
+                    deflate(patient)
+                ).to.deep.equal({ age: 35, doctor: null });
+
             });
 
             it('deserializes to null', () => {
-                const deserialized = inflate(Patient, { married: null });
-                expect(deserialized instanceof Patient).to.equal(true);
-                expect(deserialized).to.deep.equal({ married: null });
+
+                const doctor = inflate(Doctor, { name: null });
+                expect(doctor).to.be.instanceOf(Doctor);
+                expect(doctor).to.deep.equal({ name: null });
+
+                const patient = inflate(Patient, { age: 35, doctor: null });
+                expect(patient).to.be.instanceOf(Patient);
+                expect(patient).to.deep.equal({ age: 35, doctor: null });
+
             });
 
         });
@@ -105,43 +140,80 @@ describe('options behavior', () => {
 
         describe('when value is undefined and option is false/null/undefined (default)', () => {
 
+            class Doctor {
+                @Serialize() public name: string;
+            }
+
             class Patient {
-                @Serialize() public married: boolean;
-                public constructor(married?: boolean) {
-                    if (married !== undefined) { this.married = married; }
-                }
+                @Serialize() public age: number;
+                @Serialize() public doctor: Doctor;
             }
 
             it('should fail to serialize', () => {
-                const patient = new Patient(undefined);
-                expect(() => deflate(patient)).to.throw('Unable to serialize property "married": Value is undefined');
+
+                const doctor = new Doctor(); // name is undefined
+                expect(
+                    () => deflate(doctor)
+                ).to.throw('Unable to serialize property "name": Value is undefined');
+
+                const patient = Object.assign(new Patient(), { age: 35 }); // doctor is undefined
+                expect(
+                    () => deflate(patient)
+                ).to.throw('Unable to serialize property "doctor": Value is undefined');
+
             });
 
             it('should fail to deserialize', () => {
-                expect(() => inflate(Patient, { married: undefined })).to.throw('Unable to deserialize property "married": Value is undefined');
+
+                const doctorObj = {}; // name is undefined
+                expect(
+                    () => inflate(Doctor, doctorObj)
+                ).to.throw('Unable to deserialize property "name": Value is undefined');
+
+                const patientObj = { age: 35 }; // doctor is undefined
+                expect(
+                    () => inflate(Patient, patientObj)
+                ).to.throw('Unable to deserialize property "doctor": Value is undefined');
+
             });
 
         });
 
         describe('when value is undefined and option is true', () => {
 
+            class Doctor {
+                @Serialize({ optional: true }) public name: string;
+            }
+
             class Patient {
-                @Serialize({ optional: true }) public married: boolean;
-                public constructor(married?: boolean) {
-                    if (married !== undefined) { this.married = married; }
-                }
+                @Serialize({ optional: true }) public age: number;
+                @Serialize({ optional: true }) public doctor: Doctor;
             }
 
             it('serializes to undefined', () => {
-                const patient = new Patient(undefined);
-                const serialized = deflate(patient);
-                expect(serialized).to.not.haveOwnProperty('married');
+
+                const doctor = new Doctor(); // name is undefined
+                expect(
+                    deflate(doctor)
+                ).to.deep.equal({}); // name is not serialized
+
+                const patient = Object.assign(new Patient(), { age: 35 }); // doctor is undefined
+                expect(
+                    deflate(patient)
+                ).to.deep.equal({ age: 35 }); // doctor is not serialized
+
             });
 
             it('deserializes to undefined', () => {
-                const deserialized = inflate(Patient, { married: undefined });
-                expect(deserialized instanceof Patient).to.equal(true);
-                expect(deserialized).to.not.haveOwnProperty('married');
+
+                const doctor = inflate(Doctor, {}); // name is undefined
+                expect(doctor).to.be.instanceOf(Doctor);
+                expect(doctor).to.deep.equal({});
+
+                const patient = inflate(Patient, { age: 35 }); // doctor is undefined
+                expect(patient).to.be.instanceOf(Patient);
+                expect(patient).to.deep.equal({ age: 35 });
+
             });
 
         });